pyproject.toml
[tool.poetry]
name = "pizza"
version = "0.1.0"
description = ""
authors = ["Michael Napodano <michaelpnap@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
dacite = "^1.8.1"
grpcio = "^1.60.1"
grpcio-tools = "^1.60.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
